adding user dependency in coupons querys

+++ src/services/database/couponServices.js
@@ -1,21 +1,17 @@
 const Coupon = require('../../models/Coupon');
 
-exports.remove = async function (code) {
-    return await Coupon.deleteOne({ code });
+exports.find = async function (userId, length = 100, page = 1) {
+    return await Coupon.find({ user: userId }).exec();
 }
 
-exports.findAll = async function (length = 100, page = 1) {
-    return await Coupon.find({});
-}
-
-exports.find = async function (filter) {
-    return await Coupon.findOne(filter).exec();
-}
-
-exports.findByCode = async function (code) {
-    return await Coupon.findOne({ code }).exec();
+exports.findByCode = async function (code, userId) {
+    return await Coupon.findOne({ code, user: userId }).exec();
 }
 
 exports.insert = async function (coupon) {
     return await new Coupon(coupon).save();
+}
+
+exports.remove = async function (code, userId) {
+    return await Coupon.deleteOne({ code, user: userId });
 }
